# docs/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/files/upload": {
      "post": {
        "description": "Uploads a file directly to the backend, which handles S3 upload internally",
        "operationId": "FilesController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to upload"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "MEMBERSHIP_APPLICATION",
                      "BUSINESS_LICENSE",
                      "FINANCIAL_REPORT",
                      "PAYMENT_PROOF",
                      "OTHER_DOCUMENT"
                    ],
                    "description": "File category"
                  },
                  "entityId": {
                    "type": "string",
                    "description": "ID of the entity this file belongs to"
                  },
                  "entityType": {
                    "type": "string",
                    "description": "Type of entity this file belongs to"
                  },
                  "folderPath": {
                    "type": "string",
                    "description": "Target folder path for upload (e.g., /Documents/Projects)"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or file type not allowed"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Upload file directly to backend",
        "tags": [
          "Files"
        ]
      }
    },
    "/api/files/{id}": {
      "get": {
        "description": "Retrieves file metadata by ID",
        "operationId": "FilesController_getFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "File ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File information retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "File not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get file information",
        "tags": [
          "Files"
        ]
      },
      "delete": {
        "description": "Deletes the file from both storage and the database",
        "operationId": "FilesController_deleteFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "File ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully"
          },
          "404": {
            "description": "File not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete file",
        "tags": [
          "Files"
        ]
      }
    },
    "/api/files/{id}/download": {
      "get": {
        "description": "Streams the file content directly from the server",
        "operationId": "FilesController_downloadFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "File ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File content streamed",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "File not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Download file",
        "tags": [
          "Files"
        ]
      }
    },
    "/api/files/{id}/download-url": {
      "get": {
        "description": "Get download URL for the file",
        "operationId": "FilesController_downloadFileUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "File ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download URL retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "downloadUrl": {
                      "type": "string",
                      "description": "Download URL for the file"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "File not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Download file url",
        "tags": [
          "Files"
        ]
      }
    },
    "/api/files": {
      "get": {
        "description": "Retrieves all files associated with a specific entity",
        "operationId": "FilesController_getFilesByEntity",
        "parameters": [
          {
            "name": "entityType",
            "required": true,
            "in": "query",
            "description": "Type of entity (e.g., User, LeaveRequest)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": true,
            "in": "query",
            "description": "ID of the entity",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Files retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get files by entity",
        "tags": [
          "Files"
        ]
      }
    },
    "/api/files/drives/{entityType}/{entityId}/folders": {
      "post": {
        "description": "Creates a new folder in the specified drive",
        "operationId": "FilesController_createFolder",
        "parameters": [
          {
            "name": "entityType",
            "required": true,
            "in": "path",
            "description": "Type of entity",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": true,
            "in": "path",
            "description": "ID of the entity",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFolderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Folder created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "parentId": {
                      "type": "string"
                    },
                    "driveId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create folder",
        "tags": [
          "Files"
        ]
      }
    }
  },
  "info": {
    "title": "Check-in System API",
    "description": "API documentation for the Check-in System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Deploy Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "FileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "File ID"
          },
          "fileName": {
            "type": "string",
            "description": "File name"
          },
          "fileType": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO",
              "AUDIO",
              "DOCUMENT",
              "PDF",
              "EXCEL",
              "OTHER"
            ],
            "description": "File type"
          },
          "storeType": {
            "type": "string",
            "enum": [
              "LOCAL",
              "S3"
            ],
            "description": "Storage type"
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type"
          },
          "fileSize": {
            "type": "number",
            "description": "File size in bytes"
          },
          "storagePath": {
            "type": "string",
            "description": "Storage path or S3 key"
          },
          "publicUrl": {
            "type": "string",
            "description": "Public URL"
          },
          "checksum": {
            "type": "string",
            "description": "File checksum"
          },
          "entityType": {
            "type": "string",
            "description": "Entity type"
          },
          "entityId": {
            "type": "string",
            "description": "Entity ID"
          },
          "tempKey": {
            "type": "string",
            "description": "Temporary key for grouping files"
          },
          "category": {
            "type": "string",
            "description": "File category"
          },
          "description": {
            "type": "string",
            "description": "File description"
          },
          "folderId": {
            "type": "string",
            "description": "Folder ID"
          },
          "uploadedById": {
            "type": "string",
            "description": "Uploaded by user ID"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "INTERNAL"
            ],
            "description": "File visibility"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date"
          },
          "deleted": {
            "type": "boolean",
            "description": "Whether file is deleted"
          }
        },
        "required": [
          "id",
          "fileName",
          "fileType",
          "storeType",
          "storagePath",
          "visibility",
          "createdAt",
          "updatedAt",
          "deleted"
        ]
      },
      "CreateFolderDto": {
        "type": "object",
        "properties": {
          "parentPath": {
            "type": "string",
            "description": "Parent folder path"
          },
          "name": {
            "type": "string",
            "description": "Folder name"
          }
        },
        "required": [
          "parentPath",
          "name"
        ]
      }
    }
  }
}